requirements.txt
flask
ultralytics
opencv-python
pillow
gunicorn
